fix-#85zrwzc7k/fix files App.tsx and MainScreen.tsx
